1104_주기능 개발 완료

+++ MediViewer/src/main/java/com/tjoeun/mediviewer/domain/req/ReqParams.java
@@ -13,7 +13,7 @@ public class ReqParams {
 	/*
 	 * 	list request 객체
 	 */
-	private Integer studyKey;
+	private Long studyKey;
     private String pid;
     private String pname;
     private String modality;
